(#1109) Error on Patch Failures
There were a few issues in play here.
1. `yarn --frozen-lockfile` does not clean out node modules.
2. The USWDS patch can only correctly be applied to a fresh copy of
    USWDS.
3. Our postinstall command was calling lerna, which actually stripped
    out the error message from patch-package saying the patch could
    not be applied.
4. Patch-package defaults to NOT erroring when running locally, only
    in CI.
So at the end of the day, when you had a previous bootstrapped project
and you switch a branch to one with the same package.json packages,
but with an updated patch, `lerna bootstrap -- --frozen-lockfile` would
SILENTLY not apply the patch and carry on as if nothing was wrong.

This commit remediates all of the above so that bad patches fail,
errors are seen, and ncids-css modules is always cleared out so
the patches will always be applied.

Closes #1109